NEX-42 Pay Cash

Approved-by: Sasha Udovika

+++ Next2/src/Next2/Models/OrderBindableModel.cs
@@ -22,12 +22,8 @@ namespace Next2.Models
 
         public string OrderNumberText { get; set; } = string.Empty;
 
-        public double Total { get; set; }
+        public float Total { get; set; }
 
-        public double Tax { get; set; }
-
-        #region -- Private helpers --
-
-        #endregion
+        public float Tax { get; set; }
     }
 }
